+++ API-Samples/Python/ArmBased/.gitignore
@@ -1,2 +1,3 @@
 .env
 /output
+testing.py
